rename rpc-server-instance-state.h -> rpc-server-state.h
